group ref updates, pubs creation, article topics
group updates change the group ref on openPositions, researchFocuses,
techniques, followers, and members docs.

topics on articles that are added to group are added to the group,
with a rank. Rank increase with more topics, decrease with deletion
of articles.

pubs creation function is more streamlined and uses listeners
for authors, references, msauthors, and topics.